+++ exporter/opentelemetry-exporter-otlp-proto-http/src/opentelemetry/exporter/otlp/proto/http/metric_exporter/__init__.py
@@ -14,10 +14,11 @@ from __future__ import annotations
 
 import gzip
 import logging
+import random
 import zlib
 from io import BytesIO
 from os import environ
-from time import sleep
+from time import sleep, time
 from typing import (  # noqa: F401
     Any,
     Callable,
@@ -25,6 +26,7 @@ from typing import (  # noqa: F401
     Iterable,
     List,
     Mapping,
+    Optional,
     Sequence,
 )
 
@@ -33,7 +35,6 @@ from requests.exceptions import ConnectionError
 from typing_extensions import deprecated
 
 from opentelemetry.exporter.otlp.proto.common._internal import (
-    _create_exp_backoff_generator,
     _get_resource_data,
 )
 from opentelemetry.exporter.otlp.proto.common._internal.metrics_encoder import (
@@ -46,10 +47,13 @@ from opentelemetry.exporter.otlp.proto.http import (
     _OTLP_HTTP_HEADERS,
     Compression,
 )
+from opentelemetry.exporter.otlp.proto.http._common import (
+    _is_retryable,
+)
 from opentelemetry.proto.collector.metrics.v1.metrics_service_pb2 import (  # noqa: F401
     ExportMetricsServiceRequest,
 )
-from opentelemetry.proto.common.v1.common_pb2 import (  # noqa: F401  # noqa: F401
+from opentelemetry.proto.common.v1.common_pb2 import (  # noqa: F401
     AnyValue,
     ArrayValue,
     InstrumentationScope,
@@ -99,11 +103,10 @@ DEFAULT_COMPRESSION = Compression.NoCompression
 DEFAULT_ENDPOINT = "http://localhost:4318/"
 DEFAULT_METRICS_EXPORT_PATH = "v1/metrics"
 DEFAULT_TIMEOUT = 10  # in seconds
+_MAX_RETRYS = 6
 
 
 class OTLPMetricExporter(MetricExporter, OTLPMetricExporterMixin):
-    _MAX_RETRY_TIMEOUT = 64
-
     def __init__(
         self,
         endpoint: str | None = None,
@@ -111,7 +114,7 @@ class OTLPMetricExporter(MetricExporter, OTLPMetricExporterMixin):
         client_key_file: str | None = None,
         client_certificate_file: str | None = None,
         headers: dict[str, str] | None = None,
-        timeout: int | None = None,
+        timeout: float | None = None,
         compression: Compression | None = None,
         session: requests.Session | None = None,
         preferred_temporality: dict[type, AggregationTemporality]
@@ -170,7 +173,7 @@ class OTLPMetricExporter(MetricExporter, OTLPMetricExporterMixin):
         self._headers = headers or parse_env_headers(
             headers_string, liberal=True
         )
-        self._timeout = timeout or int(
+        self._timeout = timeout or float(
             environ.get(
                 OTEL_EXPORTER_OTLP_METRICS_TIMEOUT,
                 environ.get(OTEL_EXPORTER_OTLP_TIMEOUT, DEFAULT_TIMEOUT),
@@ -189,8 +192,9 @@ class OTLPMetricExporter(MetricExporter, OTLPMetricExporterMixin):
             preferred_temporality, preferred_aggregation
         )
         self._max_export_batch_size: int | None = max_export_batch_size
+        self._shutdown = False
 
-    def _export(self, serialized_data: bytes):
+    def _export(self, serialized_data: bytes, timeout_sec: float):
         data = serialized_data
         if self._compression == Compression.Gzip:
             gzip_data = BytesIO()
@@ -209,7 +213,7 @@ class OTLPMetricExporter(MetricExporter, OTLPMetricExporterMixin):
                 url=self._endpoint,
                 data=data,
                 verify=self._certificate_file,
-                timeout=self._timeout,
+                timeout=timeout_sec,
                 cert=self._client_cert,
             )
         except ConnectionError:
@@ -217,92 +221,91 @@ class OTLPMetricExporter(MetricExporter, OTLPMetricExporterMixin):
                 url=self._endpoint,
                 data=data,
                 verify=self._certificate_file,
-                timeout=self._timeout,
+                timeout=timeout_sec,
                 cert=self._client_cert,
             )
         return resp
 
-    @staticmethod
-    def _retryable(resp: requests.Response) -> bool:
-        if resp.status_code == 408:
-            return True
-        if resp.status_code >= 500 and resp.status_code <= 599:
-            return True
-        return False
-
     def export(
         self,
         metrics_data: MetricsData,
-        timeout_millis: float = 10_000,
+        timeout_millis: Optional[float] = 10000,
         **kwargs,
     ) -> MetricExportResult:
+        if self._shutdown:
+            _logger.warning("Exporter already shutdown, ignoring batch")
+            return MetricExportResult.FAILURE
+
         serialized_data = encode_metrics(metrics_data)
+        deadline_sec = time() + self._timeout
 
+        # If no batch size configured, export as single batch with retries as configured
         if self._max_export_batch_size is None:
-            for delay in _create_exp_backoff_generator(
-                max_value=self._MAX_RETRY_TIMEOUT
-            ):
-                if delay == self._MAX_RETRY_TIMEOUT:
-                    return MetricExportResult.FAILURE
-
-                resp = self._export(serialized_data.SerializeToString())
-                # pylint: disable=no-else-return
+            for retry_num in range(_MAX_RETRYS):
+                resp = self._export(
+                    serialized_data.SerializeToString(), deadline_sec - time()
+                )
                 if resp.ok:
                     return MetricExportResult.SUCCESS
-                elif self._retryable(resp):
-                    _logger.warning(
-                        "Transient error %s encountered while exporting metric batch, retrying in %ss.",
-                        resp.reason,
-                        delay,
-                    )
-                    sleep(delay)
-                    continue
-                else:
+                # multiplying by a random number between .8 and 1.2 introduces a +/20% jitter to each backoff.
+                backoff_seconds = 2**retry_num * random.uniform(0.8, 1.2)
+                if (
+                    not _is_retryable(resp)
+                    or retry_num + 1 == _MAX_RETRYS
+                    or backoff_seconds > (deadline_sec - time())
+                ):
                     _logger.error(
-                        "Failed to export batch code: %s, reason: %s",
+                        "Failed to export metrics batch code: %s, reason: %s",
                         resp.status_code,
                         resp.text,
                     )
                     return MetricExportResult.FAILURE
+                _logger.warning(
+                    "Transient error %s encountered while exporting metrics batch, retrying in %.2fs.",
+                    resp.reason,
+                    backoff_seconds,
+                )
+                sleep(backoff_seconds)
 
-            return MetricExportResult.FAILURE
-
-        # Else, attempt export in batches
+        # Else, export in batches of configured size
         split_metrics_batches = list(self._split_metrics_data(serialized_data))
         export_result = MetricExportResult.SUCCESS
 
         for split_metrics_data in split_metrics_batches:
-            # Export current batch until success, non-transient error, or timeout reached
-            for delay in _create_exp_backoff_generator(
-                max_value=self._MAX_RETRY_TIMEOUT
-            ):
-                if delay == self._MAX_RETRY_TIMEOUT:
-                    export_result = MetricExportResult.FAILURE
-                    break
-
+            # Export current batch until success, non-transient error, or retries maxed out
+            for retry_num in range(_MAX_RETRYS):
                 split_resp = self._export(
-                    split_metrics_data.SerializeToString()
+                    split_metrics_data.SerializeToString(),
+                    deadline_sec - time(),
                 )
-                # pylint: disable=no-else-return
                 if split_resp.ok:
                     export_result = MetricExportResult.SUCCESS
+                    # Move on to next batch
                     break
-                elif self._retryable(split_resp):
-                    _logger.warning(
-                        "Transient error %s encountered while exporting metric batch, retrying in %ss.",
-                        split_resp.reason,
-                        delay,
-                    )
-                    sleep(delay)
-                    continue
-                else:
+
+                # multiplying by a random number between .8 and 1.2 introduces a +/20% jitter to each backoff.
+                backoff_seconds = 2**retry_num * random.uniform(0.8, 1.2)
+                if (
+                    not _is_retryable(split_resp)
+                    or retry_num + 1 == _MAX_RETRYS
+                    or backoff_seconds > (deadline_sec - time())
+                ):
                     _logger.error(
-                        "Failed to export batch code: %s, reason: %s",
+                        "Failed to export metrics batch code: %s, reason: %s",
                         split_resp.status_code,
                         split_resp.text,
                     )
                     export_result = MetricExportResult.FAILURE
+                    # Don't retry; move on to next batch
                     break
+
+                _logger.warning(
+                    "Transient error %s encountered while exporting metric batch, retrying in %.2fs.",
+                    split_resp.reason,
+                    backoff_seconds,
+                )
+                sleep(backoff_seconds)
+                continue
 
         # Return last result after all batches are attempted
         return export_result
@@ -696,7 +699,11 @@ class OTLPMetricExporter(MetricExporter, OTLPMetricExporterMixin):
         return split_resource_metrics_pb
 
     def shutdown(self, timeout_millis: float = 30_000, **kwargs) -> None:
-        pass
+        if self._shutdown:
+            _logger.warning("Exporter already shutdown, ignoring call")
+            return
+        self._session.close()
+        self._shutdown = True
 
     @property
     def _exporting(self) -> str: